backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
yt-dlp==2023.12.30
openai==1.6.1
deepgram-sdk==3.2.0
elevenlabs==0.2.27
python-dotenv==1.0.0
requests==2.31.0
